Small improvements are done

+++ CarApp2022/src/rentACar/Application/Features/Customers/Commands/DeleteCustomerCommand.cs
@@ -2,6 +2,8 @@
 using Application.Features.Customers.Rules;
 using Application.Services.Repositories;
 using AutoMapper;
+using Core.CrossCuttingConcerns.Exceptions;
+using Core.Utilities;
 using Domain.Entities;
 using MediatR;
 
@@ -28,6 +30,10 @@ public class DeleteCustomerCommand : IRequest<DeleteCustomerDto>
         public async Task<DeleteCustomerDto> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
             await _customerBusinessRules.CheckCustomerById(request.Id);
+
+            var customerToUpdate = await _customerRepository.GetAsync(x => x.Id == request.Id);
+
+            if (customerToUpdate == null) throw new BusinessException(Messages.CustomerDoesNotExist);
 
             Customer mappedCustomer = _mapper.Map<Customer>(request);
             Customer deletedCustomer = await _customerRepository.DeleteAsync(mappedCustomer);
